+++ crates/hdk_derive/src/lib.rs
@@ -204,7 +204,7 @@ pub fn hdk_entry(attrs: TokenStream, code: TokenStream) -> TokenStream {
 }
 
 #[proc_macro_attribute]
-pub fn hdk_extern(_attrs: TokenStream, item: TokenStream) -> TokenStream {
+pub fn hdk_extern(attrs: TokenStream, item: TokenStream) -> TokenStream {
     // extern mapping is only valid for functions
     let item_fn = syn::parse_macro_input!(item as syn::ItemFn);
 
@@ -224,9 +224,17 @@ pub fn hdk_extern(_attrs: TokenStream, item: TokenStream) -> TokenStream {
 
     let internal_fn_ident = external_fn_ident.clone();
 
-    (quote::quote! {
-        map_extern!(#external_fn_ident, #internal_fn_ident, #input_type, #output_type);
-        #item_fn
-    })
-    .into()
+    if attrs.to_string() == "infallible" {
+        (quote::quote! {
+            map_extern_infallible!(#external_fn_ident, #internal_fn_ident, #input_type, #output_type);
+            #item_fn
+        })
+        .into()
+    } else {
+        (quote::quote! {
+            map_extern!(#external_fn_ident, #internal_fn_ident, #input_type, #output_type);
+            #item_fn
+        })
+        .into()
+    }
 }