+++ crates/holochain_sqlite/src/sql/cell/schedule/delete_live_ephemeral.sql
@@ -1,0 +1,5 @@
+DELETE FROM
+  ScheduledFunctions
+WHERE
+  ephemeral = TRUE
+  AND START <= ?