+++ crates/holochain_sqlite/src/sql/cell/schedule/update.sql
@@ -1,0 +1,10 @@
+UPDATE
+  ScheduledFunctions
+SET
+  maybe_schedule = :maybe_schedule,
+  START = :start,
+END = :end,
+ephemeral = :ephemeral
+WHERE
+  zome_name = :zome_name
+  AND scheduled_fn = :scheduled_fn