+++ crates/holo_hash/CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+## 0.0.7
+
+## 0.0.6
+
+### Fixed
+
+- Crate now builds with `--no-default-features`
+
 ## 0.0.5
 
 ## 0.0.4